XDR: Encode ustr and str sizes in 64 bit, RPC: test cases are working properly again

+++ common/xdrbuf.cc
@@ -403,7 +403,7 @@ namespace csl
 
     xdrbuf & xdrbuf::operator>>(common::ustr & val)
     {
-      uint32_t sz = 0;
+      uint64_t sz = 0;
       (*this) >> sz;
       uint64_t szrd=0;
